starting with main.cpp
the main.cpp file is partially filled. The vector which contains our shared pointer pointing on derived classes of Developer is already implemented. Moreover one shared pointer pointing on JuniorDeveloper as well as another shared pointer pointing on SeniorDeveloper have been implemented. The vector is already filled with the shared pointers and an iterator which goes through the vector is implemented.